clwapi/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.32.0
docker==7.1.0
pydantic==2.10.0
httpx==0.28.1
PyYAML==6.0.2
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-cov==6.0.0
